Add build-essential package before running pip install
The detection of whether pip has prebuilt wheels is not entirely
reliable.  It seems that it can be influenced by the docker host OS.
When a wheel is not found, we may need to run `gcc` and libraries to
build a native extension. `gcc` alone is not sufficient,
`build-essential` seems to be.