Re-raise errors from creating cluster in openstack
With the `abort` line the error is processed by `bad_request` defined in
`cluster_builder/__init__.py`.  Which produces an error such as the
following.  Which is a perfectly fine generic 400 error.

```
HTTP/1.1 400 BAD REQUEST
Server: Werkzeug/2.3.6 Python/3.8.17
Date: Thu, 08 Feb 2024 15:19:30 GMT
Content-Type: application/json
Content-Length: 211
Connection: close

{
  "errors": [
    {
      "detail": "ERROR: Parameter 'key_name' is invalid: Error validating value 'nope': The Key (nope) could not be found.",
      "status": "400",
      "title": "Bad Request"
    }
  ]
}
```

With the `raise e` line the error is processed by `HeatHttpBadRequestHandler`
defined in `cluster_builder/openstack/error_handling.py`.  Which
produces the following:

```
HTTP/1.1 400 BAD REQUEST
Server: Werkzeug/2.3.6 Python/3.8.17
Date: Thu, 08 Feb 2024 15:19:52 GMT
Content-Type: application/json
Content-Length: 248
Connection: close

{
  "errors": [
    {
      "detail": "Error validating value 'nope': The Key (nope) could not be found.",
      "source": {
        "pointer": "/cluster/parameters/key_name"
      },
      "status": "400",
      "title": "Bad Request"
    }
  ]
}
```

This error is largely similar, but has two differences.  The first is
the addition of the `source.pointer` value, which indicates which
user-provided value caused the issue.  Visualiser uses this value to
display the error message right next to the bad parameter.  The second
is the `detail` value, which is somewhat sanitized for presentation to
the user.