Prep for generation of cluster types
The cluster types that we wish to support differ in three regards:

* The user data passed to the components.
* Which components constitute them.
* The number of components, e.g., compute nodes, that constitute them.

Which components constitute a cluster type has already been addressed.
The cluster type author can specify which components are mandatory and
can also allow the end-user to select from some optional components.

The number of components has still to be completed.

This commit addresses passing different user data to the same component
for different cluster types.  That is the `gateway` component requires
different user data for a slurm cluster than it does for a kubernetes
cluster.

The user data has been removed from the component HOTs.  They now
reference a user data file at a relative path.  As a relative path is
used and the component HOTs are identical except for the user data,
there is no need to generate the HOT, instead it can placed at a path
relative to the desired user data file.

Generation of the user data files, as opposed to embeded user data as a
YAML formatted string, is simpler as it doesn't need to concern itself
with significant indentation, whilst still remaining readable.