Adjust how params and groups are defined
Instead of defining parameters in parameters.yaml they are now defined
in the components.  This allows for selecting which parameters to
include based on the selected components allow for parameters on
optional compoents that don't have a default.

The parameter_groups are now defined cluster-type.yaml.  This makes
documenting the purpose of `optional.name` a little easier and removes
some complexity from loading a cluster type.