Build artefacts for all heat backed cluster types
The components and user data files for the SLURM: Team Edition and
Container Cruncher (Medium) cluster types have been generated with the
build script and committed to git.

There are no changes to the SLURM: Team Edition components or user data
files because the generated files are identical to the hand-crafted
files.