First pass at configurable number of nodes
The `nodes` component now has a `OS::Heat::ResourceGroup` which
references the `node` component.  The `node` component is a single
instance of what was in the `nodes` component previously.

We can use the `hardcoded_parameters` mechanism to, err, hard code the
number of nodes.  Or we could allow the user to specify the number.