JupyterLab Jumpstart cluster type
This requires the introduction of a new component `gateway-standalone`.
It differs from the `gateway` component in the following ways:

* It does not have a shared storage volume.
* It takes a parameter for the size of the gateways non-shared volume.
* It sets the password for the flight user from a parameter.

The reason for the difference in behaviour around setting the password
for the flight user is that we are targetting two different users with
the standalone and team environments.

The team environment is targeted at a user with some admin/ops
experience.  They will be comfortable with using passwordless SSH to
access the cluster.  They will setup users using IPA and disseminate
those login credentials as they see fit.

The standalone (single user) environment is targeted at a user who
has very limited, if any, admin/ops experience.  They are expected to
access the cluster via Fligh Web Suite.  If the password were set as for
the team environment, they would have to SSH into the cluster and then
either reset the password or understand how to retrieve the generated
password.  Better to ask them to provide it when they launch the
cluster.