Generate a single cluster.yaml component
For non-optional components, generate a single cluster.yaml file instead
of one-per component.  For our current cluster types, this results in
either a single `cluster.yaml` component or `cluster.yaml` and
`node.yaml` components.

This change makes it easier to test the generated components by
launching them directly to HEAT. It also makes it clearer what the
cluster consists of.