Support per-cluster type parameter overrides
Each cluster type can now specify optional parameter overrides.  When
the `cluster.yaml` (and other components) are generated, their
parameters are updated with any overrides specified for that cluster
type.

One use case for this might be to have a generic `node-count` parameter,
whcih is then overridden by a small, medium and large cluster type.  We
previously supported that as long as the parameter was hardcoded.  Now
we also provide the option of adding additional constraints and changing
the defaults.

Another use case, might be to add additional constraints on which
flavours are available for a give parameters.  For instance `m1.tiny`
might not be suitable for running a Flight Solo image.

Any value that is overridden is replaced entirely, so if overriding the
constraints, care needs to be taken to include all desired constraints.