Add instructions to cluster types
Cluster types have gained instructions on their launch, administration
and usage.  These have been largely copied from the instructions on
Flight Warehouse.

Visualiser will display the launch instructions on the cluster launch
page and the admin and usage instructions on the rack page. In order to
have the rack display the instructions for the correct cluster type, it
is necessary to have the stack annotated with the cluster type that
created it in some manner.  A `cluster-type` parameter has been added to
the `gateway` and `gateway-standalone` components and each appropriate
cluster type updated to hardcode that parameter.

A better solution would be to have visualiser create a cluster model
when it launches the cluster and add any bookkeeping records itself once
the launch request has been accepted.  However given current
conversations around concertim.com I suspect that that change is better
left until then.